A WebTiles client library for Python
This is a very simple library can be use to make clients that interact
with DCSS WebTiles servers. It's asynchronous and built on the modules
asyncio and websockets. Its primary focus at the moment is to provide
support for the [beem](https://github.com/gammafunk/beem) project.

Initially the library has a base WebTilesConnection class that
supports connecting over the websocket, authentication, parsing of the
game list, and updating the RC file, a WebTilesLobbyConnection class
inheriting from this that adds parsing lobby messages, and an
incomplete WebTilesGameConnection class. I may merge
WebTilesLobbyConnection with WebTilesConnection later, as the former
only handles a few extra json message types, but
WebTilesGameConnection will probably stay separate, at least least in
terms of game interaction. WebTilesGameConnection supports the variety
of functions needed for watching games and interacting with chat on a
basic level. It may be simpler to have WebTilesGameConnection support
basic game operations and use something else in a submodule to handle
game-interaction like parsing game data and sending game commands.

There's a full working example in updaterc.py, which is installed as the
script update-dcss-rc. This script is taken from the
[dcss-rc](https://github.com/gammafunk/dcss-rc) project where it
previously had its own implementation of a simple WebTiles client. It
now correctly handles tasks without deleting websockets before they're
closed and allows specifying servers both through a server code and a
full websocket url. The set of known servers has been updated to remove
cszo and add cjr.