Give some properties better names and update documentation
The `game_username` property wasn't the best name, since it might not
actually be the username of the player, and `username` property was too
generic. Now game_username is watch_username and username is
login_username, to avoid any confusion. We also no longer store the
password after connecting, since it's not needed anywhere.