Preliminary support for the the webtiles-changes protocol
This supports logging in and parsing lobby and chat messages under the
"version 2" JSON protocol used on the webtiles-changes branch. Game list
parsing isn't yet available. There's a protocol_version argument to
connect() and documentation is updated where relevant.

This commit moves the parse_chat() function used by beem into the
WebTilesGameConnection class as parse_chat_message() and uses the html
library to unescape HTML entities, instead of doing this through a
hard-coded list of substitutions.

This also bumps the version to 0.2.0.