Add game parsing support for webtiles-changes and update update-dcss-rc
Now games in the V2 protocol are parsed in the games property of
WebTilesConnection after login, just as for the V1 protocol.

For update-dcss-rc, it now supports updates webtiles-changes
servers. The default servers have their protocols specified and CPO is
added to this list. The command-line syntax allows a vN+url syntax to
specify an optional protocol version and websocket url in one
argument. For instance, an explicit command-line specification of CPO
would be:

v2+wss://crawl.project357.org/socket

The other command-line syntax is unchanged.