Add a get_rc method WebTilesConnection
Sends the RC retrieval request to the server. Probably not much use for
this on most servers, where RC files are available through e.g. Apache,
but now implemented as part of the protocol all the same. Clients will
have to respond to "rcfile_contents" messages after sending this, as
noted in the docs.